Test code cleanup
