+++ test/files.js
@@ -1,0 +1,22 @@
+const Files = {
+  DEFAULT: './test/res/default.epub',
+  UNZIPPED_DEFAULT: './test/res/default',
+  EXPECTED_DEFAULT_CONTEXT: './test/res/expectedDefaultContext.json',
+  EXPECTED_DEFAULT_BOOK: './test/res/expectedDefaultBook.json',
+  INVALID_PACKAGE: './test/res/invalidPackage.epub',
+  INVALID_XML: './test/res/invalidXml.epub',
+  NCX_MISSING: './test/res/ncxMissing.epub',
+  META_INF_MISSING: './test/res/metainfMissing.epub',
+  OPF_MISSING: './test/res/opfMissing.epub',
+  EXTRACT_STYLE: './test/res/extractStyle.epub',
+  UNZIPPED_EXTRACT_STYLE: './test/res/extractStyle',
+  EXPECTED_EXTRACT_STYLE_CONTEXT: './test/res/expectedExtractStyleContext.json',
+  EXPECTED_EXTRACT_STYLE_BOOK: './test/res/expectedExtractStyleBook.json',
+  EXPECTED_EXTRACT_STYLES: './test/res/expectedExtractStyles.json',
+  EXPECTED_EXTRACT_STYLES_WITH_BASE_PATH: './test/res/expectedExtractStylesWithBasePath.json',
+  EXPECTED_EXTRACT_BODY: './test/res/expectedExtractBody.json',
+  EXPECTED_EXTRACT_BODY_WITH_NO_ADAPTOR: './test/res/expectedExtractBodyWithNoAdaptor.json',
+  EXPECTED_READ_SPINE_WITH_BASE_PATH: './test/res/expectedReadSpineWithBasePath.xhtml',
+};
+
+export default Files;
